r3_74: merged the HEAD WvStreams and corresponding changes from r3_73.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/r3_74@2637 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ linuxstreams/wvfam.cc
@@ -5,9 +5,14 @@
 #include "wvistreamlist.h"
 #include <sys/stat.h>
 
-WvFamBase::~WvFamBase()
+void WvFamBase::close()
 {
+    if (!s)
+        return;
+
     WvIStreamList::globallist.unlink(s);
+    delete s;
+    s = 0;
 
     if (FAMClose(&fc) == -1)
         log(WvLog::Error, "%s\n", FamErrlist[FAMErrno]);
@@ -59,14 +64,14 @@ void WvFamBase::_callback()
                 || fe.code == FAMCreated)
         {
             if (!fe.userdata)
-                cb(WvString(fe.filename), WvFamEvent(fe.code));
+                cb(WvString(fe.filename), WvFamEvent(fe.code), false);
 
             // If the void * points to something this is a directory callback.
             // We'll prepend the path to the returned filename.
             else
                 cb(WvString("%s/%s",
                     *reinterpret_cast<WvString *>(fe.userdata),
-                    fe.filename), WvFamEvent(fe.code));
+                    fe.filename), WvFamEvent(fe.code), true);
         }
     }
 
@@ -105,7 +110,7 @@ void WvFamBase::setup()
 
         s->setcallback(WvStreamCallback(this, &WvFamBase::callback), 0);
 
-        WvIStreamList::globallist.append(s, true);
+        WvIStreamList::globallist.append(s, false);
     }
 }
